backend/requirements.txt
google-python-cloud-debugger

Django==2.2.13
django-cors-headers==3.1.1

djangorestframework==3.10.3

# Postgresql
psycopg2==2.8.4

# authentication
django-allauth==0.40.0

django-phonenumber-field==2.2.0
phonenumbers==8.10.5

GoogleAppEngineCloudStorageClient==1.9.15.0
django-google-cloud-storage==0.0.2
google-cloud-storage==1.14.0
Pillow==6.2.1

pandas==0.25.1

gspread==3.1.0
oauth2client==1.5.2
pyOpenSSL==19.0.0

google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib

plotly==4.6.0

django-admin-sortable2==0.7.6
django-taggit==1.2.0
django-autoslug==1.9.7

python-frontmatter==0.5.0

django-extensions==2.2.9
django-filter==2.2.0
django_countries==6.1.3

mock

google-cloud-error-reporting